renaming remaining poll to pool
