+++ src/grpc-clients/src/mqtt/placement/mod.rs
@@ -21,7 +21,7 @@ use protocol::broker_mqtt::broker_mqtt_placement::mqtt_broker_placement_service_
 use tonic::transport::Channel;
 
 use super::MqttBrokerPlacementInterface;
-use crate::poll::ClientPool;
+use crate::pool::ClientPool;
 
 pub mod call;
 pub mod inner;
@@ -38,7 +38,7 @@ async fn placement_client(
 
 pub(crate) async fn placement_interface_call(
     interface: MqttBrokerPlacementInterface,
-    client_poll: Arc<ClientPool>,
+    client_pool: Arc<ClientPool>,
     addr: String,
     request: Vec<u8>,
 ) -> Result<Vec<u8>, CommonError> {
