+++ src/mqtt-broker/src/handler/session.rs
@@ -136,7 +136,7 @@ mod test {
     use std::sync::Arc;
 
     use common_base::config::broker_mqtt::BrokerMqttConfig;
-    use grpc_clients::poll::ClientPool;
+    use grpc_clients::pool::ClientPool;
     use metadata_struct::mqtt::session::MqttSession;
     use protocol::mqtt::common::ConnectProperties;
